Create churn_idd_by_gw_permonth.py
This is how you can filter out the churn from your DB based on their appeared during the last few months. If they made a purchase or any other transaction. The Gw can be considered as a category. So you may look for a churn by category instead of a total churn.